IGN altitude service

+++ src/main/java/com/mass/flightplan/vac/AirportInfo.java
@@ -12,10 +12,19 @@ import java.util.List;
 @ToString
 public class AirportInfo {
     @NonNull String code, name, eAipVersion;
-    @NonNull int altitude;
+
+    /**
+     * Altitude in meters AMSL
+     */
+    @NonNull double altitude;
+
     @NonNull double latitude;
+
     @NonNull double longitude;
+
     @NonNull int localPressure;
+
     @NonNull String contactInfo;
+
     @NonNull List<RunwayInfo> runways;
 }
